VoteRealBackend/requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
beautifulsoup4==4.11.2
bs4==0.0.1
cachetools==5.3.0
cdp-backend==4.0.5
cdp-scrapers==0.6.1
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_fdy66ybhuv/croot/certifi_1669749572520/work/certifi
cffi==1.15.1
charset-normalizer==3.0.1
civic-scraper==0.2.5
clean-text==0.6.0
click==8.1.3
click-option-group==0.5.5
cryptography==39.0.1
dataclasses-json==0.5.7
decorator==5.1.1
defusedxml==0.7.1
demjson3==3.0.6
docx2txt==0.8
emoji==1.7.0
esprima==4.0.1
feedparser==6.0.10
filelock==3.9.0
fireo==1.7.0
flatbuffers==23.1.21
frozenlist==1.3.3
fsspec==2023.1.0
ftfy==6.1.1
gast==0.4.0
gcsfs==2023.1.0
google-api-core==2.11.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.3
grpcio-status==1.51.3
h5py==3.8.0
huggingface-hub==0.12.1
icalendar==5.0.4
idna==3.4
jax==0.4.4
keras==2.12.0rc1
libclang==15.0.6.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.23.5
oauthlib==3.2.2
openai==0.26.5
opt-einsum==3.3.0
packaging==23.0
pdfminer.six==20221105
pdfplumber==0.8.0
Pillow==9.4.0
proto-plus==1.22.2
protobuf==4.22.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
scrapelib==2.1.0
scraper-legistar==0.0.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.12.0rc0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
tokenizers==0.13.2
tqdm==4.64.1
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.14
Wand==0.6.11
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.15.0
yarl==1.8.2